fix unit test single conv
the addresses dont match exactly the POC because now that we include
the node's weight in the sum of neighbors weight, in many cases the neighborWeights
is the same (all nodes are connected), and so the choice between those are random.
There's no importance to which of them gets allocated first.