Fix unit tests
Now that tensor_graph_coloring uses the right size of the CMX, some unit tests
were failing since they run a conv that doesn't fit in a CMX slice, and
the tests disabled MaxTopologicalCut, so tensor graph coloring is failing because
the layer cannot fit in CMX slice.

When disabling MaxTopologicalCut pass, best if TensorGraphColoring and
serialize pass be disabled, since if tensor graph coloring is not run
the addresses are not assigned and serialization will fail.