VPUNND-1156: Add barriers to the computational graph in one pass
Previously the opModel was iterated through first, followed
by the controlModel, which caused some barriers to not
appear in order after serialization when weights prefetching
was turned on.

Remove this by iterating over a topologically sorted
control model and inserting barriers due to both control
and data flows for a given node in one pass of the control model
graph.

Signed-off-by: Meenakshi Venkataraman <meenakshi.venkataraman@intel.com>