Graph Coloring: fix aggressive simplify to make it faster
- making a copy of TIG for very large graphs is very slow, so replacing the
method of making a copy and work on that to decide the order of coloring, by reusing the
existing graph.
The copy is replaced with:
-  updating each node with a flag identifying if it has already been selected by aggressive simplify
-  marking neighbors weight with a value that will force it to be ordered last
by the ordering strategy algorithm

This approach now limits the Ordering Strategy methods to only
IG_LARGEST_NEIGHBORS_FIRST or IG_SMALLEST_NEIGHBORS_FIRST
(we are using IG_LARGEST_NEIGHBORS_FIRST right now) the other strategies (with the exception of random)
can also be supported by the above, with adding some more fields to the node.