Hack to use always the concatOfConcats rather than implicitJoin
The implicitJoin option has issues when next layer needs to be streamed
since this pass happens before GO, the decision on whether we can
use ImplicitJoin approach is not yet well formed.

- cons of using concatToConcat approach is that we always go to DDR after, even
if next layer can run in CMX, but that doesnt for sure mean worse performance since
with implicitJoin we need to DMA storageElements and sparsity for next layer.