Adapt barrier scheduler for mutliple output UPATasks
Networks with multiple outputs where each input has a chain of
UPATasks can end up with redundant wait dependencies after the
removal of a barrier.

This happens with the vehicle-attributes-recognition-barrier-0042
model which has two Softmax outputs. After introducing Conversion
layers after each Softmax, the scheduler ends up removing the
barrier between the Softmax and Conversion layers, while leaving
one of the Conversions with a dependency on the barrier.

This commit takes both layers into account so that the dependency
can be properly removed.