[MLIR] Add special Dialect for lazy constant folding
The `Const` Dialect provides utilities for that lazy constant folding support.

It defines a special attribute `Const::ContentAttr` to hold original constant data
and the transformations applied to it.
The `Const::ContentAttr` provides an API to apply the transformations on-the-fly,
when accessing the data.

The `Const` Dialect supports the following transformations:

* Precision conversion
* Quantization types casting from raw storage
* Dequantization
* Reshape
* Reorder
* Padding
* SubView

Replace old mechnism with the new Dialect usage.